+++ lib/has_states/stateable.rb
@@ -5,14 +5,15 @@ module HasStates
     extend ActiveSupport::Concern
 
     included do
-      has_many :states, class_name: 'HasStates::State',
-                        as: :stateable,
-                        dependent: :destroy
+      has_many :states, class_name: 'HasStates::Base',
+                       as: :stateable,
+                       dependent: :destroy
     end
 
     # Instance methods for managing states
-    def add_state(type, status: 'pending', metadata: {})
+    def add_state(type, status: 'pending', metadata: {}, state_class: HasStates::State)
       states.create!(
+        type: state_class.name,
         state_type: type,
         status: status,
         metadata: metadata