+++ spec/generators/tmp/db/migrate/20241223213845_create_has_states_states.rb
@@ -1,6 +1,7 @@
  class CreateHasStatesStates < ActiveRecord::Migration[8.0]
   def change
     create_table :has_states_states do |t|
+      t.string :type, null: false
       t.string :state_type
       t.string :status, null: false
 
@@ -8,9 +9,9 @@
 
       t.references :stateable, polymorphic: true, null: false
 
-      t.datetime :completed_at
       t.timestamps
 
+      t.index %i[type stateable_id]
       t.index %i[stateable_type stateable_id]
     end
   end